scripts/pythonactors/requirements.txt
setproctitle
paho-mqtt

# anticipated
pyyaml>=5.4
